feat(apie): Convert interactive playground to SSG ✅
PHASE 1 - CONVERSION 2/4 COMPLETE

Changes:
- Extract all interactive logic to ApiePlaygroundClient.tsx (415 lines)
- Create server component page.tsx with force-static flag
- Add proper SEO metadata for playground activities
- Keep ALL functionality: audio, confetti, coloring, jokes

Results:
✅ Build successful - no RSC/hook errors
✅ Static HTML generated (.next/server/app/apie.html - 16.8kB)
✅ SEO meta: 'Play with Apie in our interactive playground!'
✅ Bundle size stable (141kB - expected for Phase 1)

Complex interactive playground now statically rendered! 🎨🎵
Next: /accepted page conversion
